security review: fix settlement overflow (i128) + honest threat model
Phase 9.1 self-review (docs/SECURITY.md).
- FIXED: integer overflow in settle_deltas — qty×price could exceed i64 and wrap
  silently in release. Now i128 end-to-end (engine + service + sim); service applies
  deltas to u64 balances clamped to [0, u64::MAX]. Conservation re-tested (9/9).
- DOCUMENTED (honest caveats needing production hardening): no operation
  authentication (cancel/withdraw aren't signed — single-operator trust model),
  work-item replay (no per-op nonce), collateralization/underflow clamp,
  commit-reveal griefing + builder trust.
- SOUND (and why): determinism, clearing optimality, settlement conservation incl.
  fees, market integrity, MEV-resistance — each property-tested. "Trustless" scoped
  precisely to the audited matching/settlement.

+++ crates/match-engine/src/lib.rs
@@ -243,8 +243,8 @@ mod tests {
             if let Some((price, entries)) = wire::decode_settlement(&blob) {
                 // conservation holds for any fee, with the treasury included in the sum
                 let deltas = wire::settle_deltas(price, &entries, fee_bps, u32::MAX);
-                let sum_base: i64 = deltas.iter().map(|d| d.1).sum();
-                let sum_quote: i64 = deltas.iter().map(|d| d.2).sum();
+                let sum_base: i128 = deltas.iter().map(|d| d.1).sum();
+                let sum_quote: i128 = deltas.iter().map(|d| d.2).sum();
                 prop_assert_eq!(sum_base, 0);
                 prop_assert_eq!(sum_quote, 0);
             }
